feat: finish invoice workflow

+++ ms_invoice_management/domain/model/src/main/java/co/com/management/model/ProductInfo.java
@@ -2,7 +2,7 @@ package co.com.management.model;
 
 public record  ProductInfo(String[] names,
                            Integer[] quantities,
-                           Float[] prices,
+                           Double[] prices,
                            String[] codes
 ) {
 }
