+++ ms_invoice_management/domain/model/src/main/java/co/com/management/model/product/Product.java
@@ -1,11 +1,10 @@
 package co.com.management.model.product;
-import lombok.Builder;
+
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.util.UUID;
 
 @Getter
 @Setter
@@ -16,5 +15,5 @@ public class Product {
     private String code;
     private String name;
     private Integer quantity;
-    private Float unitPrice;
+    private Double unitPrice;
 }
